moves input sanitation tests to x-objective endpoints.
