+++ tests/Endpoints/ObjectiveTest.php
@@ -142,87 +142,4 @@ class ObjectiveTest extends ReadWriteEndpointTest
             'fullCoursesThroughSession' => [[1, 2], ['fullCourses' => [1]]],
         ];
     }
-
-    /**
-     * Ideally, we'd be testing the "purified textarea" form type by itself.
-     * However, the framework currently does not provide boilerplate to roll container-aware form test.
-     * We'd need a hybrid between <code>KernelTestCase</code> and <code>TypeTestCase</code>.
-     * @link  http://symfony.com/doc/current/cookbook/testing/doctrine.html
-     * @link http://symfony.com/doc/current/cookbook/form/unit_testing.html
-     * To keep things easy, I bolted this test on to this controller test for the time being.
-     * @todo Revisit occasionally and check if future versions of Symfony have addressed this need. [ST 2015/10/19]
-     *
-     * @dataProvider inputSanitationTestProvider
-     *
-     * @param string $input A given objective title as un-sanitized input.
-     * @param string $output The expected sanitized objective title output as returned from the server.
-     *
-     */
-    public function testInputSanitation($input, $output)
-    {
-        $postData = $this->getContainer()->get(ObjectiveData::class)
-            ->create();
-        $postData['title'] = $input;
-        unset($postData['id']);
-
-        $this->createJsonRequest(
-            'POST',
-            $this->getUrl($this->kernelBrowser, 'app_api_objectives_post', [
-                'version' => $this->apiVersion
-            ]),
-            json_encode(['objectives' => [$postData]]),
-            $this->getAuthenticatedUserToken($this->kernelBrowser)
-        );
-
-        $response = $this->kernelBrowser->getResponse();
-
-        $this->assertJsonResponse($response, Response::HTTP_CREATED);
-        $this->assertEquals(
-            json_decode($response->getContent(), true)['objectives'][0]['title'],
-            $output,
-            $response->getContent()
-        );
-    }
-
-    /**
-     * @return array
-     */
-    public function inputSanitationTestProvider()
-    {
-        return [
-            ['foo', 'foo'],
-            ['<p>foo</p>', '<p>foo</p>'],
-            ['<ul><li>foo</li></ul>', '<ul><li>foo</li></ul>'],
-            ['<script>alert("hello");</script><p>foo</p>', '<p>foo</p>'],
-            [
-                '<a href="https://iliosproject.org" target="_blank">Ilios</a>',
-                '<a href="https://iliosproject.org" target="_blank" rel="noreferrer noopener">Ilios</a>'
-            ],
-        ];
-    }
-
-    /**
-     * Assert that a POST request fails if form validation fails due to input sanitation.
-     */
-    public function testInputSanitationFailure()
-    {
-        $postData = $this->getContainer()->get(ObjectiveData::class)
-            ->create();
-        // this markup will get stripped out, leaving a blank string as input.
-        // which in turn will cause the form validation to fail.
-        $postData['title'] = '<iframe></iframe>';
-        unset($postData['id']);
-
-        $this->createJsonRequest(
-            'POST',
-            $this->getUrl($this->kernelBrowser, 'app_api_objectives_post', [
-                'version' => $this->apiVersion
-            ]),
-            json_encode(['objectives' => [$postData]]),
-            $this->getAuthenticatedUserToken($this->kernelBrowser)
-        );
-
-        $response = $this->kernelBrowser->getResponse();
-        $this->assertJsonResponse($response, Response::HTTP_BAD_REQUEST);
-    }
 }
